+++ app/src/main/java/com/battlelancer/seriesguide/util/ClipboardTools.kt
@@ -34,9 +34,20 @@ private val onLongClickListener = View.OnLongClickListener {
     return@OnLongClickListener it is TextView && copyTextToClipboard(it.context, it.text)
 }
 
+private val onContextClickListener = View.OnContextClickListener {
+    return@OnContextClickListener it is TextView && copyTextToClipboard(it.context, it.text)
+}
+
+/**
+ * Sets a long click listener and on Android 6 or newer a context click listener that enables
+ * right clicks with a mouse.
+ */
 fun TextView.copyTextToClipboardOnLongClick() {
-    // globally shared click listener instance
+    // globally shared click listener instances
     setOnLongClickListener(onLongClickListener)
+    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+        setOnContextClickListener(onContextClickListener)
+    }
 }
 
 fun View.copyTextToClipboardOnLongClick(text: CharSequence) {