+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Version 2024.4
 * 🔧 Calendar: add discoverable more options button on items, does the same thing as touch and hold.
 * 🔧 Display more options of list items on touch and hold and right click (for example in Samsung
   Dex) as well.
+* 🔧 Where touch and hold copies text, support right click to copy as well.
 
 ### 2024.4.3 🧪
 *2024-09-13*